fix circular
