+++ services/workflows-service/src/app.module.ts
@@ -14,7 +14,8 @@ import { StorageModule } from './storage/storage.module';
 import { MulterModule } from '@nestjs/platform-express';
 import { EventEmitterModule } from '@nestjs/event-emitter';
 import { FilterModule } from '@/filter/filter.module';
-import { configs, env, validate } from '@/env';
+import { configs, env } from '@/env';
+import { validate } from '@/env-validate';
 import { SentryModule } from '@/sentry/sentry.module';
 import { RequestIdMiddleware } from '@/common/middlewares/request-id.middleware';
 import { AxiosRequestErrorInterceptor } from '@/common/interceptors/axios-request-error.interceptor';
