+++ services/workflows-service/src/app.worker.module.ts
@@ -4,7 +4,8 @@ import { ClsModule } from 'nestjs-cls';
 
 import { AnalyticsModule } from '@/common/analytics-logger/analytics.module';
 import { AppLoggerModule } from '@/common/app-logger/app-logger.module';
-import { configs, env, validate } from '@/env';
+import { configs, env } from '@/env';
+import { validate } from '@/env-validate';
 import { SecretsManagerModule } from '@/secrets-manager/secrets-manager.module';
 import { SentryModule } from '@/sentry/sentry.module';
 import { WebhooksModule } from '@/webhooks/webhooks.module';
